# ENH: where for ufunc reductions