… doc/release/1.17.0-notes.rst
… numpy/core/tests/test_umath.py