Note: This is a simpler and more robust, yet faster, version of #12635 and #12640 (way too much time wasted on being clever, but the comments helped steer to this better solution).

This introduces a `where` keyword for reductions. It works well, but for ufuncs with no identity, one has to explicitly pass in an `initial`. If people agree this is the way to go, I will add documentation and more tests.

Unlike for my other attempts, this performs quite well and I think we should consider using it in `nanfunctions` and in `MaskedArray`:
```
a = np.arange(100000.)
a[1000] = np.nan
np.add.reduce(a, where=~np.isnan(a)) == np.nansum(a)
# True
%timeit np.nansum(a)
# 10000 loops, best of 5: 113 µs per loop
%timeit np.add.reduce(a, where=~np.isnan(a))
# 10000 loops, best of 5: 92.3 µs per loop

# Now try what will be the worst-case scenario.
a[::2] = np.nan
np.add.reduce(a, where=~np.isnan(a)) == np.nansum(a)
# True
%timeit np.nansum(a)
# 1000 loops, best of 5: 381 µs per loop
%timeit np.add.reduce(a, where=~np.isnan(a))
# 1000 loops, best of 5: 389 µs per loop
```