@njsmith 2019-01-02
Neat!

> for ufuncs with no identity, one has to explicitly pass in an `initial`. 

Can you explain why?