@seberg 2019-01-02
I think I may rather want to skip the `initial` part to leave the option of adding `identity` specifically.