@mhvk 2019-01-02
@shoyer - `initial` was initially called `identity` but after discussion was replaced only by `initial` - see #10635. In our current reductions, it is actually a better description of what happens...

@njsmith - the reason for having to pass in `initial` is the same as in my initial attempt, to be able to deal with the case where no elements are selected (but here, it is truly an initial, not an identity, as it is no longer used as a replacement).