@seberg 2019-01-02
@mhvk sorry, I think I interpreted it the wrong way around. I though `initial` was assumed to be an identity for the operation here.