@mhvk 2019-01-02
OK, now with initial comments addressed and more tests and docs.