@njsmith 2019-01-02
So if you have a reduction that has no identity, and you try to use `where=` but don't pass `initial=`, then is that always an error, or is it only an error when one of the reductions turns out to have zero elements?