@mhvk 2019-01-02
Right now, without an identity and initial, using `where` always errors. To change this would mean quite a bit of hacking in the code that determines the initial from the data itself: currently, it simply takes a view of the first element of the input data along the reduction axis, but obviously that won't work if that first element might be masked.  I would like to postpone this to later (or not do it at all; not sure it is worth it...).