@eric-wieser 2019-01-03
> not sure it is worth it

Without doing it, I don't think we can use it to implement things like summary masked arrays of dtype=object