@mhvk 2019-01-03
> Without doing it, I don't think we can use it to implement things like summary masked arrays of dtype=object

I don't quite follow. Currently, the `MaskedArray.sum` just fills the masked pieces with `0`, so it is assumed that the object array knows what to do when summing with `0`. It would seem that nothing would break if one were to use `initial=0`.

p.s. For object arrays one could pass in a special `initial` object that has `def self.__add__(other): return other`, and then at the end replace any remaining instances of those with `masked`, thus avoiding the need to reduce the mask as well. But I don't see how to generalize this to non-object.