@mhvk 2019-01-03
@eric-wieser - on `MaskedArray.sum`, I tried replacing `self.filled(0).sum(...)` with `self._data.sum(..., where=~_mask, initial=0)` and almost all masked array tests pass. The exception are tests with masked matrices, because I did not update `matrix.sum` to recognize `where` (or `initial` for that matter). Which I guess brings up a general problem of using this in `MaskedArray` - it would need to be able to fall back to using `filled` for a while to support other masked subclasses, while emitting a `FutureWarning` or so.