@mhvk 2019-01-03
p.s. I guess the first use of this may be in `nanfunctions`...