@eric-wieser 2019-01-03
> For object arrays one could pass in a special initial object that has def self.__add__(other): return other

That's a really neat idea!