@eric-wieser 2019-01-03
Other than the reference leak above, this looks good to go for me