@mhvk 2019-01-03
OK, fixed the reference issue by re-using `_wheremask_converter`, and rebased/squashed since it seems ready to go in.