@ahaldane 2019-01-03
Also, one case I thought about a lot has to do with non-commutative reductions. For instance:
```python
>>> np.equal.reduce([False, False], initial=False)                            
False
>>> np.equal.reduce([False, False, False], initial=False)
True
```
if you work it out, you'll see that the end result depends on how successive values alternate. Therefore, there is a difference between replacing a value by an identity element before reducing (my strategy), and not using the element in the reduction in the first place (this PR, correct?).