@eric-wieser 2019-01-03
> if you work it out, you'll see that the end result depends on how successive values alternat

Right, but `np.equal` has no identity. For the special case of the `??->?` loop, it has an identity of `True`, for which your above would work just fine.