@mhvk 2019-01-03
@ahaldane - yes, I think this should work well for the general masking case. The one problem I noted [earlier](https://github.com/numpy/numpy/pull/12644#issuecomment-451179117) is that whatever is being masked has to be able to deal with `where` and `initial` in their own `sum()` method. But if a project is all up to date with the numpy API, it should be fine.

p.s. Will look at your `MaskedArray` class; curious to see if it would allow astropy finally to have a `MaskedQuantity`...