@mattip 2019-01-04
Nice and simple enhancement. Would be nice to add the comparison benchmark at some point