@seberg 2019-01-04
Yeah, looked good to me as well. Only thing to maybe quickly follow up on: Do we want to allow allow non-boolean masks? I think I would prefer to not allow it for now but the code seems to me like it does, and there should be a test for it probably.