@seberg 2019-01-04
Sorry, nevermind, the loop uses "safe" casting, thought I skimmed an unsafe cast there somewhere.