@mhvk 2019-01-04
> Nice and simple enhancement.  Would be nice to add the comparison benchmark at some point

I added a new issue #12662 to use this in `nanfunctions` - item 1 for that is to add benchmarks for those functions.