@mhvk 2019-01-04
@seberg - mask array creation and casting is now identical to what is used for the regular ufuncs, so OK to the extent those are OK!