@ahaldane 2019-01-04
I added a commit using it in the MaskedArray ducktype, and it passes my test suite and gives a fair speedup. Setup:
```python                                       
>>> np.random.seed(12345)                                                      
>>> x = np.random.rand(515, 512)                                                
>>> a = MaskedArray(x, x < 0.5) 
```

Before this PR:
```python
>>> %timeit np.add.reduce(a, axis=1)    
1.63 ms ± 8.96 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
After this PR:
```python                                      
>>> %timeit np.add.reduce(a, axis=1)                                           
1.39 ms ± 2.37 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

The case I tested might be a worst-case for this PR, too, since it has lots of alternating masked elements.