@mhvk 2019-01-04
@ahaldane - happy to see that it helps! And, yes, that'll definitely be close to worst-case performance.