@ahaldane 2019-01-04
Apologies, I mixed myself up on the timings. Here is the correct comparison:

Before:
```python
>>> a = MaskedArray(x, x < 0.9)                                                
>>> %timeit np.add.reduce(a, axis=1)                                           
751 µs ± 3.55 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
>>> a = MaskedArray(x, x < 0.5)
>>> %timeit np.add.reduce(a, axis=1)      
1.4 ms ± 3.07 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
>>> a = MaskedArray(x, x < 0.1)                                               
>>> %timeit np.add.reduce(a, axis=1)                                          
748 µs ± 16 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

After:
```python
>>> a = MaskedArray(x, x < 0.9)                                                
>>> %timeit np.add.reduce(a, axis=1)                                           
600 µs ± 1.97 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)    
>>> a = MaskedArray(x, x < 0.5)
>>> %timeit np.add.reduce(a, axis=1)      
1.63 ms ± 3.13 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)                       
>>> a = MaskedArray(x, x < 0.1)                                               
>>> %timeit np.add.reduce(a, axis=1)                                          
684 µs ± 15.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
So if I did the timings right it seems be faster in the two biased cases, but slower in the worst case.